📐 🍱 Update metric key data structure (#759)
This PR is cherry picked from #744 since that PR has gotten a bit unruly. This PR allows for one part of it to be independently integrated without really changing anything about how PyKEEN works
