+++ src/pykeen/evaluation/rank_based_evaluator.py
@@ -6,10 +6,9 @@ import itertools as itt
 import logging
 import math
 import random
-import re
 from collections import defaultdict
 from dataclasses import dataclass, field, fields
-from typing import DefaultDict, Dict, Iterable, List, Mapping, NamedTuple, Optional, Sequence, Tuple, Union
+from typing import DefaultDict, Dict, Iterable, List, Mapping, Optional, Sequence, Tuple, Union
 
 import numpy as np
 import pandas as pd
@@ -18,6 +17,19 @@ from dataclasses_json import dataclass_json
 from scipy import stats
 
 from .evaluator import Evaluator, MetricResults, prepare_filter_triples
+from .metrics import (
+    ADJUSTED_ARITHMETIC_MEAN_RANK,
+    ADJUSTED_ARITHMETIC_MEAN_RANK_INDEX,
+    ARITHMETIC_MEAN_RANK,
+    GEOMETRIC_MEAN_RANK,
+    HARMONIC_MEAN_RANK,
+    INVERSE_ARITHMETIC_MEAN_RANK,
+    INVERSE_GEOMETRIC_MEAN_RANK,
+    INVERSE_HARMONIC_MEAN_RANK,
+    INVERSE_MEDIAN_RANK,
+    MEDIAN_RANK,
+    MetricKey,
+)
 from .ranks import Ranks
 from ..triples.triples_factory import CoreTriplesFactory
 from ..typing import (
@@ -25,9 +37,9 @@ from ..typing import (
     LABEL_HEAD,
     LABEL_RELATION,
     LABEL_TAIL,
-    RANK_REALISTIC,
-    RANK_TYPE_SYNONYMS,
     RANK_TYPES,
+    SIDE_BOTH,
+    SIDES,
     MappedTriples,
     Target,
 )
@@ -36,23 +48,9 @@ from ..utils import fix_dataclass_init_docs
 __all__ = [
     "RankBasedEvaluator",
     "RankBasedMetricResults",
-    "MetricKey",
-    "resolve_metric_name",
 ]
 
 logger = logging.getLogger(__name__)
-
-SIDE_BOTH = "both"
-SIDES = {LABEL_HEAD, LABEL_TAIL, SIDE_BOTH}
-
-ARITHMETIC_MEAN_RANK = "arithmetic_mean_rank"  # also known as mean rank (MR)
-GEOMETRIC_MEAN_RANK = "geometric_mean_rank"
-HARMONIC_MEAN_RANK = "harmonic_mean_rank"
-MEDIAN_RANK = "median_rank"
-INVERSE_ARITHMETIC_MEAN_RANK = "inverse_arithmetic_mean_rank"
-INVERSE_GEOMETRIC_MEAN_RANK = "inverse_geometric_mean_rank"
-INVERSE_HARMONIC_MEAN_RANK = "inverse_harmonic_mean_rank"  # also known as mean reciprocal rank (MRR)
-INVERSE_MEDIAN_RANK = "inverse_median_rank"
 
 RANK_STD = "rank_std"
 RANK_VARIANCE = "rank_var"
@@ -74,100 +72,6 @@ all_type_funcs = {
     RANK_MAD: stats.median_abs_deviation,
     RANK_COUNT: lambda x: np.asarray(x.size),
 }
-
-ADJUSTED_ARITHMETIC_MEAN_RANK = "adjusted_arithmetic_mean_rank"
-ADJUSTED_ARITHMETIC_MEAN_RANK_INDEX = "adjusted_arithmetic_mean_rank_index"
-TYPES_REALISTIC_ONLY = {ADJUSTED_ARITHMETIC_MEAN_RANK, ADJUSTED_ARITHMETIC_MEAN_RANK_INDEX}
-
-METRIC_SYNONYMS = {
-    "adjusted_mean_rank": ADJUSTED_ARITHMETIC_MEAN_RANK,
-    "adjusted_mean_rank_index": ADJUSTED_ARITHMETIC_MEAN_RANK_INDEX,
-    "amr": ADJUSTED_ARITHMETIC_MEAN_RANK,
-    "aamr": ADJUSTED_ARITHMETIC_MEAN_RANK,
-    "amri": ADJUSTED_ARITHMETIC_MEAN_RANK_INDEX,
-    "aamri": ADJUSTED_ARITHMETIC_MEAN_RANK_INDEX,
-    "igmr": INVERSE_GEOMETRIC_MEAN_RANK,
-    "iamr": INVERSE_ARITHMETIC_MEAN_RANK,
-    "mr": ARITHMETIC_MEAN_RANK,
-    "mean_rank": ARITHMETIC_MEAN_RANK,
-    "mrr": INVERSE_HARMONIC_MEAN_RANK,
-    "mean_reciprocal_rank": INVERSE_HARMONIC_MEAN_RANK,
-}
-
-
-class MetricKey(NamedTuple):
-    """A key for the kind of metric to resolve."""
-
-    name: str
-    side: str
-    rank_type: str
-    k: Optional[int]
-
-    def __str__(self) -> str:  # noqa: D105
-        components = [self.name, self.side, self.rank_type]
-        if self.k:
-            components.append(str(self.k))
-        return ".".join(components)
-
-
-_SIDE_PATTERN = "|".join(SIDES)
-_TYPE_PATTERN = "|".join(itt.chain(RANK_TYPES, RANK_TYPE_SYNONYMS.keys()))
-METRIC_PATTERN = re.compile(
-    rf"(?P<name>[\w@]+)(\.(?P<side>{_SIDE_PATTERN}))?(\.(?P<type>{_TYPE_PATTERN}))?(\.(?P<k>\d+))?",
-)
-HITS_PATTERN = re.compile(r"(hits_at_|hits@|h@)(?P<k>\d+)")
-
-
-def resolve_metric_name(name: str) -> MetricKey:
-    """Functional metric name normalization."""
-    match = METRIC_PATTERN.match(name)
-    if not match:
-        raise ValueError(f"Invalid metric name: {name}")
-    k: Union[None, str, int]
-    name, side, rank_type, k = [match.group(key) for key in ("name", "side", "type", "k")]
-
-    # normalize metric name
-    if not name:
-        raise ValueError("A metric name must be provided.")
-    # handle spaces and case
-    name = name.lower().replace(" ", "_")
-
-    # special case for hits_at_k
-    match = HITS_PATTERN.match(name)
-    if match:
-        name = "hits_at_k"
-        k = match.group("k")
-    if name == "hits_at_k":
-        if k is None:
-            k = 10
-        # TODO: Fractional?
-        try:
-            k = int(k)
-        except ValueError as error:
-            raise ValueError(f"Invalid k={k} for hits_at_k") from error
-        if k < 0:
-            raise ValueError(f"For hits_at_k, you must provide a positive value of k, but found {k}.")
-    assert k is None or isinstance(k, int)
-
-    # synonym normalization
-    name = METRIC_SYNONYMS.get(name, name)
-
-    # normalize side
-    side = side or SIDE_BOTH
-    side = side.lower()
-    if side not in SIDES:
-        raise ValueError(f"Invalid side: {side}. Allowed are {SIDES}.")
-
-    # normalize rank type
-    rank_type = rank_type or RANK_REALISTIC
-    rank_type = rank_type.lower()
-    rank_type = RANK_TYPE_SYNONYMS.get(rank_type, rank_type)
-    if rank_type not in RANK_TYPES:
-        raise ValueError(f"Invalid rank type: {rank_type}. Allowed are {RANK_TYPES}.")
-    elif rank_type != RANK_REALISTIC and name in TYPES_REALISTIC_ONLY:
-        raise ValueError(f"Invalid rank type for {name}: {rank_type}. Allowed type: {RANK_REALISTIC}")
-
-    return MetricKey(name, side, rank_type, k)
 
 
 @fix_dataclass_init_docs
@@ -361,11 +265,13 @@ class RankBasedMetricResults(MetricResults):
 
         >>> metric_results.get('hits@5')
         """
-        metric, side, rank_type, k = resolve_metric_name(name)
-        if not metric.startswith("hits"):
-            return getattr(self, metric)[side][rank_type]
-        assert k is not None
-        return self.hits_at_k[side][rank_type][k]
+        return self._get_metric(MetricKey.lookup(name))
+
+    def _get_metric(self, metric_key: MetricKey) -> float:
+        if not metric_key.name.startswith("hits"):
+            return getattr(self, metric_key.name)[metric_key.side][metric_key.rank_type]
+        assert metric_key.k is not None
+        return self.hits_at_k[metric_key.side][metric_key.rank_type][metric_key.k]
 
     def to_flat_dict(self):  # noqa: D102
         return {f"{side}.{rank_type}.{metric_name}": value for side, rank_type, metric_name, value in self._iter_rows()}
