+++ tests/mocks.py
@@ -8,9 +8,8 @@ import torch
 from torch import nn
 
 from pykeen.evaluation import Evaluator, MetricResults, RankBasedMetricResults
-from pykeen.evaluation.rank_based_evaluator import SIDES
 from pykeen.nn.emb import RepresentationModule
-from pykeen.typing import RANK_REALISTIC, RANK_TYPES, MappedTriples
+from pykeen.typing import RANK_REALISTIC, RANK_TYPES, SIDES, MappedTriples
 
 __all__ = [
     "CustomRepresentations",
